requirements.txt
streamlit==1.34.0
pandas==2.2.1
transformers==4.39.3
torch==2.2.2
boto3==1.34.93
botocore==1.34.93
python-dotenv==1.0.1